updated admin, added emailing and role changing

+++ app/Controllers/Http/CommentLikeController.js
@@ -6,6 +6,7 @@
 
 /** @type {typeof import('@adonisjs/../../app/Models/CommentLike')} */
 const Like = use('App/Models/CommentLike')
+const NotFoundException = use('App/Exceptions/NotFoundException')
 
 /**
  * Resourceful controller for interacting with commentlikes
@@ -19,17 +20,13 @@ class CommentLikeController {
    * @param {Request} ctx.request
    * @param {Response} ctx.response
    */
-   async like ({ request, response, session, auth }) {
-    const {comment_id} = request.all()
+  async like({ params, response, session, auth }) {
+    const { id } = params
     const like = new Like();
-    try{
-      like.comment_id = comment_id
-      like.user_id = auth.user.id
-      await like.save()
-      session.flash({success: 'Liking successful'})
-    } catch(e) {
-      session.flash({error: 'Liking unsuccessful'})
-    }
+    like.comment_id = id
+    like.user_id = auth.user.id
+    await like.save()
+    session.flash({ success: 'Liking successful' })
     response.redirect('back')
   }
 
@@ -41,15 +38,12 @@ class CommentLikeController {
    * @param {Request} ctx.request
    * @param {Response} ctx.response
    */
-  async dislike ({ params, request, response }) {
-    const {id} = params
-    try{
-      const like = await Like.find(id)
-      await like.delete()
-      session.flash({success: 'Unliking successful'})
-    } catch(e) {
-      session.flash({success: 'Unliking unsuccessful'})
-    }
+  async dislike({ params, request, response, session }) {
+    const { id } = params
+    const like = await Like.findBy('comment_id', id)
+    if (!like) throw new NotFoundException()
+    await like.delete()
+    session.flash({ success: 'Unliking successful' })
     response.redirect('back');
   }
 }
